==> 2_From multiple csv files to one csv file.ipynb <==
This notebook is used to combine csv files into a larger csv file, because it is also easier to work with the data when it is aggregated into one file.

import os
import pandas as pd
from pathlib import Path

os.chdir(r'.\kb_csv_files') # move to folder with the many csv files

# collect all csv filtered in Danish language
files = os.listdir()
files_dan = [i for i in files if 'KB_metadata_dan_' in i]

list_of_dan_df = [pd.read_csv(i) for i in files_dan]
dan_df = pd.concat(list_of_dan_df).reset_index(drop=True)

#############################

# collect all csv not filtered in language danish
other_files = [i for i in files if 'KB_metadata_dan_' not in i]

list_of_other_dfs = [pd.read_csv(i) for i in other_files]
other_df = pd.concat(list_of_other_dfs).reset_index(drop=True)


# Save dataset
# move to the folder where you save the large datasets
os.chdir('..')
os.chdir('kb_metadata_set')

# Save
dan_df.to_csv('kb_metadata_set_dan_raw.csv', index=False)
other_df.to_csv('kb_metadata_set_all_raw.csv', index=False)
