==> 3_Filter and clean.ipynb <==
This notebook is an example on how to filter and clean the data.

import pandas as pd
import os

os.chdir(r'.\kb_metadata_set')

### Filter rows using str.contains

# Load data
dan_df = pd.read_csv('kb_metadata_set_dan_raw.csv')


# filter column "sub 3" on value "code o"
dan_df_filtered = dan_df[dan_df['sub_3'].str.contains('code o', na=False)].reset_index(drop=True)


# clear the column
L = list(dan_df_filtered['sub_3'])
new_l = [' '.join(i.split(',')[1:]).replace(")",'')\
         .replace("'",'')\
         .replace(" -",',-')\
         .replace('The special pressure collection', '').strip() for i in L]

# Store cleaned data in a new column; sub_3_st
dan_df_filtered['sub_3_st'] = new_l

### Extract years using str.extract and regex, and saving values in a new column; 'year_st'

dan_df_filtered['year_st'] = dan_df_filtered['year'].str.extract(r'(\d{4})')

### Clean up author values
There are two "author" columns. I collect the values ​​in a column and clean the text strings.

# Collect the values ​​in two author columns in pairs
dan_df_filtered['author_st'] = list(zip(dan_df_filtered['author1'], dan_df_filtered['author2']))

def clean_author_1_2(text):
 text = str(text)
 text_replace = text.replace('(', '').replace(')', '').replace('nan,', '').replace(', nan', '').replace("' ", '')
 text_strip = text_replace.strip()
 return text_strip

dan_df_filtered['author_st'] = dan_df_filtered['author_st'].apply(lambda x : clean_author_1_2(x))

os.chdir('..')
os.chdir('kb_filtered_datasets')

### Select a few of the columns for a new dataframe

You can reorder and select the columns that you wish.

df_filtered = dan_df_filtered[['author_st', 'title', 'year_st', 'publisher', 'place', 'sub_4', 'sub_3_st', 'misc_comments']]

# save the clean and filtered data
df_filtered.to_csv('kb_metadata_dan_filtered.csv', index=False)
